rearranged the SDK files to split things up - needs shared fixtures for the tests still
